Get App Ready to Deploy

+++ app/src/main/java/org/henryschmale/milespergallontracker/MileageEvent.java
@@ -30,7 +30,7 @@ public class MileageEvent {
 
     @ColumnInfo(name = "car_id")
     @NonNull
-    public int carId;
+    public long carId;
 
     @NonNull
     public long mileage;
